+++ src/backend/ci/core/auth/api-auth/src/main/kotlin/com/tencent/devops/auth/pojo/UserPermissionInfo.kt
@@ -27,8 +27,6 @@
 
 package com.tencent.devops.auth.pojo
 
-import com.tencent.devops.common.auth.api.AuthPermission
-import com.tencent.devops.common.auth.api.AuthResourceType
 import io.swagger.annotations.ApiModel
 import io.swagger.annotations.ApiModelProperty
 
@@ -38,6 +36,6 @@ data class UserPermissionInfo(
     val organizationId: Int,
     @ApiModelProperty("组织级别")
     val organizationLevel: Int,
-    @ApiModelProperty("权限详情")
-    val permissionMap: Map<AuthResourceType, List<AuthPermission>>
+    @ApiModelProperty("权限详情 key:资源类型 value: action集合")
+    val permissionMap: Map<String, List<String>>
 )
