+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/resources/op/OpManagerUserResourceImpl.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.auth.resources
+package com.tencent.devops.auth.resources.op
 
 import com.tencent.devops.auth.api.manager.OpManagerUserResource
 import com.tencent.devops.auth.pojo.ManagerUserEntity
