+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/resources/user/UserProjectMemberResourceImpl.kt
@@ -26,7 +26,7 @@
  *
  */
 
-package com.tencent.devops.auth.resources
+package com.tencent.devops.auth.resources.user
 
 import com.tencent.bk.sdk.iam.constants.ManagerScopesEnum
 import com.tencent.bk.sdk.iam.dto.manager.ManagerRoleGroupInfo
