+++ src/backend/ci/core/auth/biz-auth/src/main/kotlin/com/tencent/devops/auth/resources/user/UserProjectRoleResourceImpl.kt
@@ -26,7 +26,7 @@
  *
  */
 
-package com.tencent.devops.auth.resources
+package com.tencent.devops.auth.resources.user
 
 import com.tencent.devops.auth.api.user.UserProjectRoleResource
 import com.tencent.devops.auth.pojo.DefaultGroup
